Update app-numeric.R
This web app is an Iris Predictor that deploys a ML model based off the Iris dataset. 
This  app allows the user to select the numeric input values for the 4 input parameters with the aim of predicting the class of an Iris plant.
See the pdf for more information regarding the Iris dataset itself.

This app uses the Shiny Web App framework.